Include checkstyle. Relocate AttempResult class.
