+++ homeworks/HW2/src/main/java/App.java
@@ -1,3 +1,4 @@
+import ru.vershinin.homework2.AttempResult;
 import ru.vershinin.homework2.Game;
 import ru.vershinin.homework2.WrongLengthExeption;
 
@@ -7,10 +8,11 @@ import java.util.Scanner;
  * This Java source file was generated by the Gradle 'init' task.
  */
 public class App {
-    private static final org.slf4j.Logger log = org.slf4j.LoggerFactory.getLogger(App.class);
-    private static Scanner in = new Scanner(System.in);
 
-    public String getGreeting() {
+    private static final org.slf4j.Logger log = org.slf4j.LoggerFactory.getLogger(App.class);
+    private static final Scanner in = new Scanner(System.in);
+
+    String getGreeting() {
         return "Hello world.";
     }
 
@@ -19,7 +21,7 @@ public class App {
         startNewGame();
     }
 
-    public static void startNewGame() {
+    private static void startNewGame() {
 
         Game game = Game.getNewGame();
 
@@ -29,22 +31,19 @@ public class App {
 
         log.info(game.getSecretWord());
 
-        for (int i = 0; i < attempCount; i++ ) {
+        for (int i = 0; i < attempCount; i++) {
             String word = in.next();
 
             try {
-                Game.AttempResult attempResult = game.attempt(word);
+                AttempResult attempResult = game.attempt(word);
 
                 if (attempResult.getResult()) {
                     System.out.println("You won!");
                     break;
-                }
-                else
-                {
+                } else {
                     System.out.println("Bulls " + attempResult.getBulls());
                     System.out.println("Cows " + attempResult.getCows());
                 }
-
             } catch (WrongLengthExeption wrongLengthExeption) {
                 System.out.println("Bad length of the word!");
             }
@@ -56,10 +55,9 @@ public class App {
         System.out.println("Wanna play again? Y/N");
         String answer = in.next();
 
-        if ( answer.indexOf('Y') != -1 || answer.indexOf('y') != -1 ) {
+        if (answer.indexOf('Y') != -1 || answer.indexOf('y') != -1) {
             startNewGame();
-        }
-        else {
+        } else {
             System.out.println("Have a nice day!");
         }
     }
